requirements.txt
pytest
string
math
random
matplotlib.pyplot